+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### master
 - restore pane layout for each window
+- bugfix: correct pane ordering in a window
 
 ### v0.0.3, 2014-08-26
 - save and restore current and alternate session