+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ### master
+
+### v0.0.4, 2014-08-26
 - restore pane layout for each window
 - bugfix: correct pane ordering in a window